netconf/xml: index should be decimal type
In netconf xml file, index node as a key value for sja1105 yang model.
The yang model set 'index' as int32 type since need to different the
entrys.

xml file elements value only could be decimal value if it is integer type.
"0x" ahead only treated as string, could not be as integer.

Signed-off-by: Po Liu <Po.Liu@nxp.com>