removed const attribute in struct sja1105_spi_setup
The attribute const has been removed for the fields device and
staging_area because these two fields can point to dynamically
allocated memory, too. This memory then needs to be freed leading to
the situation that a const pointer will be freed. Now the memory is always
allocated dynamically and thus always correctly freed. This increases
robustness and portability to other platforms, e.g. when running on LS1046.

Signed-off-by: Kay Potthoff <Kay.Potthoff@microsys.de>