tool: Skip tables that are not understood on 'config load'
SGMII is only available for the R and S devices. But currently some
versions of the sja1105-tool insert the sgmii-table and some don't, even
if the table is empty. So at least be compatible with XMLs generated by
the programs which do add this empty table (by ignoring it), even though
an SGMII configuration will not be applied.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>